% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{cds_df}
\alias{cds_df}
\title{A cds_df constructor function from a tibble}
\usage{
cds_df(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Possibly to export at a later stage, but now is used to maintain best
practices.
}
