% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_downloading.R
\name{cds_filter_fun}
\alias{cds_filter_fun}
\title{Conveniently include CDS Web API filter functions in download_cds()}
\usage{
cds_filter_fun(cds_fun_name, PropertyName = NULL, ...)
}
\arguments{
\item{cds_fun_name}{A string representing the name of the filter function.
Available functions: \href{https://docs.microsoft.com/en-us/dynamics365/customer-engagement/web-api/queryfunctions?view=dynamics-ce-odata-9}{Web API Query Function Reference}}

\item{PropertyName}{Most filter functions work on at least one field from the
base entity in your query. This argument allows you to specify this column as
an R string. If a function does not use any fields from an entity, leave is as
NULL (default value).}

\item{...}{If a query function uses more arguments than just PropertyName,
you can specify them here. As argument names use the same names as you can
find in the documentation.}
}
\value{
A string that represents CDS Web API Odata call that can be used as
argument in the Odata $filter system function. See usage examples.
}
\description{
CDS Web API offers a variety of filtering functions that are not standard
in Odata. They are very useful when it comes to limiting the number of rows
you might want to retrieve from an entity. The problem is that writing
them in a query is annoying to say at least. This function will generate a
string that callse the specified filter function wirhout need to encapsulate
strings in ticks or remembering how arrays of values are written in odata.
See list of available filter functions here:
\href{https://docs.microsoft.com/en-us/dynamics365/customer-engagement/web-api/queryfunctions?view=dynamics-ce-odata-9}{Web API Query Function Reference}
}
\examples{
cds_filter_fun("On", PropertyName = "createdon", PropertyValue = "2020-02-21")
cds_filter_fun("Between", "sale", PropertyValues = c(100, 300))
\dontrun{
download_cds(
  "audits",
  filter = cds_filter_fun("LastXFiscalYears", "createdon", PropertyValue = 2)
)
}

}
