% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_manipulation.R
\name{create_cds_records}
\alias{create_cds_records}
\title{Easily created CDS records from a tibble/ data frame.}
\usage{
create_cds_records(
  cds_df_like,
  cds_instance_name,
  target_entity_set_name = NULL,
  exclude_null_cols = FALSE
)
}
\arguments{
\item{cds_df_like}{tiible or data,frame with columns named as fields in the
target entity name and rows represeinting individula records to be created.}

\item{cds_instance_name}{A character scalar used in
\code{initialize_cds_token()} to name a CDS instance.}

\item{target_entity_set_name}{A plural name of the entity to which you want
to post new records.}

\item{exclude_null_cols}{A logical scalar indicating whether columns with
nulls should be included in a single record PATCH request. Especially useful
if you have null lookup field in your data for certain records and you want
to exclude just the to prevent the whole request from failing.}
}
\value{
List of httr requests objects for each record created. Useful to
investigate whether all requests went through and repeat for those who did
not.
}
\description{
This function will send chained POST requests to target entity in order to
create new records with columns of a cds_df_like. For the moment this is a
simple function, does not perform any data validation, putting all strain
of validation checks on your CDS instance.
}
