% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_downloading.R
\name{download_cds}
\alias{download_cds}
\title{Download data from CDS as a formatted tibble with metadata.}
\usage{
download_cds(
  query,
  cds_instance_name = NULL,
  select = NULL,
  filter = NULL,
  order_by = NULL,
  top = NULL
)
}
\arguments{
\item{query}{A character scalar with the odata request for which you wish
to download a result. Typically it will be a plural logical name of an entity
("contacts", "accounts", "phonecalls", etc.). It is passed directly in a
request so complex queries with odata operators are possible. The only
condition is that it has to perform a retrieve operation on a main entity.}

\item{cds_instance_name}{A character scalar used in
\code{initialize_cds_token()} to name a CDS instance. Default NULL. If only
one cds_token object is present in \code{getOption("cdsR.tokens")}, then you
do not need to specify this argument.}

\item{select}{A string or vector of strings with names of attributes you wish
to select from an entity. If  a vector, all values will be collapsed
using "," as separator.}

\item{filter}{A string or vector of strings with odata filter statements.
If vector, all values will be collapsed using " and " simmilarly to dplyr's
filter.}

\item{order_by}{A string or vector of strings expressing entity fields
by which you wish to arrange the result. If  a vector, all values will be
collapsed using "," as separator.}

\item{top}{A scalar integer determining how many rows of the base entity you
wish to retrieve.}
}
\value{
A cds_df S3 object, which is essentially a tibble with additional
attributes described in the \emph{Details} section.
}
\description{
This is the main function of the package which allows for easy downloading
of data from CDS into your R session. It provides the CDS metadata via
attributes "md_ent" and "md_attributes" as tibbles as well, so you can
benefit from it when automatizing the analytics tasks.
}
\details{
Tibbles returned by this function have several properties
\itemize{
  \item All enum type fields (picklists, status and state) are returned as
  proper character vectors containing labels, not the integer keys that
  the Web API normally returns.
  \item In "md_ent" attribute (\code{attr(a_cds_df, "md_ent")}) you can find
  all entity metadata listed \href{https://docs.microsoft.com/en-us/dynamics365/customer-engagement/web-api/entitymetadata?view=dynamics-ce-odata-9}{here}.
  \item In "md_attributes" attribute (\code{attr(a_cds_df, "md_attributes")})
  you can find all details regarding returned entitiy's attributes. The
  full list can be found \href{https://docs.microsoft.com/en-us/dynamics365/customer-engagement/web-api/attributemetadata?view=dynamics-ce-odata-9}{here}.
  \item In "md_enums_dicts" attribute
  (\code{attr(a_cds_df, "md_enums_dicts")}) you can find list of tibbles
  representing key, value pairs for enums fields which were used to parse
  them
}
}
