% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{dplyr_drops_attributes}
\alias{dplyr_drops_attributes}
\alias{mutate.cds_df}
\alias{inner_join.cds_df}
\alias{left_join.cds_df}
\alias{right_join.cds_df}
\alias{full_join.cds_df}
\alias{nest_join.cds_df}
\title{Use dplyr's join and mutate functions wihtout loosing CDS metadata attributes}
\usage{
\method{mutate}{cds_df}(cds_df, ...)

\method{inner_join}{cds_df}(cds_df, ...)

\method{left_join}{cds_df}(cds_df, ...)

\method{right_join}{cds_df}(cds_df, ...)

\method{full_join}{cds_df}(cds_df, ...)

\method{nest_join}{cds_df}(cds_df, ...)
}
\arguments{
\item{cds_df}{A cds_df tibble.}

\item{...}{Arguments passed to dplyr mutate and join functions and their
variants.}
}
\description{
This function is an implementation of dplyr's mutate and join generic
functions for cds_df S3 object that are returned from `downloand_cds()`.
You don't have to worry abount anything and just use dplyr normally.
}
\details{
This is needed due
to fact that `dplyr::mutate()` and joint do not preserve non-tibble
attributes. More information can be found
\href{https://github.com/tidyverse/dplyr/issues/1984}{here}.
}
