% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{evaluate_attrs}
\alias{evaluate_attrs}
\title{Evaluate a function without loosing CDS metadata attributes.}
\usage{
evaluate_attrs(cds_df, dplyr_f, ...)
}
\arguments{
\item{cds_df}{A cds_df tibble.}

\item{dplyr_f}{A function.}

\item{...}{arguments passed to `dplyr_F`}
}
\value{
Output of the `dplyr_f` function evaluated with `...` with CDS
metadata attributes preserved.
}
\description{
Used primarly to implement S3 generics coming from dplyr which remove
CDS attributes and are popular (mainly mutate, and join variants).
}
