% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{get_cds_md}
\alias{get_cds_md}
\title{Get CDS metadata from cds_df tibbles.}
\usage{
get_cds_md(cds_df, which_md = "entity")
}
\arguments{
\item{cds_df}{A cds_df tibble, typically downloaded via
\code{cdsr::download_cds()}.}

\item{which_md}{A character scalar indicating what kind of CDS metadata you
would like to get from the \code{cds_df}. Must be one of the following:
"entity", "attributes" or "picklists". The default is "entity".}
}
\value{
A tibble, or list of tibbles in case of "picklists", containing the
requested CDS metadata information.
}
\description{
This function is a convienient getter function for accessing the attributes
of cds_df objects which contain CDS metadata for the entity from which a
cds_df was downloaded, as well as attributes and picklists. Use this
function instead of \code{base::attr()} or \code{base::attributes()}.
}
