% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{is.cds_df}
\alias{is.cds_df}
\title{Check if a tibble is a cds_df}
\usage{
is.cds_df(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A boolean scalar.
}
\description{
Checks whether a tibble is a cds_df downloaded via `download_cds()`.
}
