% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{parse_cds_enums}
\alias{parse_cds_enums}
\title{Parse enum type columns (picklists, states, etc.) to keys or labels.}
\usage{
parse_cds_enums(cds_df, to = "key")
}
\arguments{
\item{cds_df}{A cds_df tibble downloaded via `download_cds()`.}

\item{to}{A character scalar, either 'key' or 'label', defining what you
want to parse into. Default is 'key'.}
}
\value{
A cds_df with all present enum columns parsed either to it's key or
label, depending on the `to` argument.
}
\description{
This function can be used to easily swap the labels or keys of enum fields
present in a cds_df. Especially useful when launching custom requests and
not using `update_cds_records()`.
}
