% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_df.R
\name{type_cds_df}
\alias{type_cds_df}
\title{Parse cds_df data into native R data types.}
\usage{
type_cds_df(cds_df)
}
\arguments{
\item{cds_df}{A cds_df to be typed.}
}
\description{
Had to be separeted from cds_df as otherwise it was caousing infinite loop
with dplyr's mutate. Also can be expanded for additional options.
}
